MNPlusOneDBSpanDetector
