+++ src/sentry/utils/performance_issues/detectors/__init__.py
@@ -1,6 +1,7 @@
 from .consecutive_db_detector import ConsecutiveDBSpanDetector  # NOQA
 from .consecutive_http_detector import ConsecutiveHTTPSpanDetector  # NOQA
 from .file_io_main_thread_detector import FileIOMainThreadDetector  # NOQA
+from .mn_plus_one_db_span_detector import MNPlusOneDBSpanDetector  # NOQA
 from .n_plus_one_api_calls_detector import NPlusOneAPICallsDetector  # NOQA
 from .n_plus_one_db_span_detector import (  # NOQA
     NPlusOneDBSpanDetector,
