+++ tests/sentry/utils/performance_issues/test_m_n_plus_one_db_detector.py
@@ -9,8 +9,8 @@ from sentry.models.options.project_option import ProjectOption
 from sentry.testutils import TestCase
 from sentry.testutils.performance_issues.event_generators import get_event
 from sentry.testutils.silo import region_silo_test
+from sentry.utils.performance_issues.detectors import MNPlusOneDBSpanDetector
 from sentry.utils.performance_issues.performance_detection import (
-    MNPlusOneDBSpanDetector,
     PerformanceProblem,
     _detect_performance_problems,
     get_detection_settings,
